refactor: delete leaky pass-through barrels and dead re-exports
Five recurring instances of the same CLAUDE.md "no convenience
barrels" violation, none documenting a public surface and most already
bypassed by their real call sites:

- src/agent/storage/index.ts: drop the 8-symbol executionLease
  re-export block. Most callers already reached past it to
  '@agent/storage/executionLease' directly for symbols the barrel
  didn't curate; the remainder now do too. Adds
  '@agent/storage/executionLease' as a new direct CLI-host specifier,
  so config/ratchets/host-agent-import-baseline.json is updated to
  match (only bare '@agent/storage' was previously listed there, and
  it stays in active CLI use via 8 other files).
- src/shared/schemas/index.ts: drop the dead `mainViewMessages`
  namespace alias (zero consumers; `export * from './mainView'` right
  above it is the real, used export).
- packages/desktop/src/desktopCommandSurface.ts: drop the SETTINGS_TAB
  and formatDesktopAccelerator re-exports; the one real consumer
  (renderer/main.ts) now imports them from their origin modules, same
  as main/desktopShellIpc.ts already did for SETTINGS_TAB.
- packages/extension/src/{commands,frontend}/*/index.ts: delete the 9
  no-README barrels under commands/{agent,auth,files,latex,system,
  setup,taskFormState} and frontend/{agents,files}; every caller now
  imports the defining file directly.

Test-side fixes for mocks/imports that pointed at the deleted barrels
(RunAgentOwnership, ExecutionRegistry, WorkflowScriptTool,
DesktopCommandSurface, MainViewMessageHandler, ExecutionHandlers,
register vitest suites) are included so the mocked module matches
where the production code now actually imports from. The
RunExecution.vitest.mts "already lost" lease test now imports
ExecutionLeaseLostError dynamically alongside the module under test,
since an earlier test's vi.resetModules() would otherwise leave its
static top-level binding pointing at a stale module instance and fail
the `instanceof` check.

No behavior change: every deletion is a real symbol/file ceasing to
exist, with callers rewritten to the file that already defines it.
