+++ config/ratchets/host-agent-import-baseline.json
@@ -30,6 +30,7 @@
       "@agent/runtime/terminalResultToast",
       "@agent/storage",
       "@agent/storage/conversationFormat",
+      "@agent/storage/executionLease",
       "@agent/trace",
       "@agent/utils/agentConfigToTaskState",
       "@agent/utils/outputFileUtils"
