--- packages/extension/src/commands/agent/index.ts
@@ -1,7 +1,0 @@
-// Barrel export for agent commands
-export { stopAgent, compactResponse } from './agentCommands';
-export { handleCreateAgentWithAI } from './agentCreatorCommands';
-export { runExecuteCommand } from './executeCommand';
-export { registerFollowUpCommand } from './followUpCommand';
-export { registerMergeCommands } from './mergeCommands';
-export { registerResumeAgentCommand } from './resumeCommand';
