--- packages/extension/src/commands/auth/index.ts
@@ -1,3 +1,0 @@
-// Re-export AUTH_COMMANDS for external use
-export { AUTH_COMMANDS } from '@auth/constants';
-export { getAuthStatus, signIn, signOut } from './authCommands';
