+++ packages/extension/src/commands/files/fileSelectionCommands.ts
@@ -4,7 +4,8 @@ import * as vscode from 'vscode';
 // Local imports - log
 import { registerCommands } from '@commands/_shared/registerCommands';
 import { getFilterExtensions } from '@common/files';
-import { FILE_SELECTION_COMMAND_IDS, getFileLister } from '@frontend/files';
+import { getFileLister } from '@frontend/files/fileLister';
+import { FILE_SELECTION_COMMAND_IDS } from '@frontend/files/fileSelectionRegistry';
 import { showLoggedErrorMessage } from '@frontend/ui/errorHandlingUtils';
 import { selectFile, selectFiles } from '@frontend/ui/dialogs';
 import * as logger from '@logger/logUtils';
