--- packages/extension/src/commands/files/index.ts
@@ -1,3 +1,0 @@
-// Barrel export for file commands
-export { registerFileSelectionCommands } from './fileSelectionCommands';
-export { registerOpenFileCommands } from './openFileCommands';
