--- packages/extension/src/commands/latex/index.ts
@@ -1,3 +1,0 @@
-// Barrel export for latex commands
-export { arXivCommands, downloadArXivSource } from './arXivCommands';
-export { registerLatexdiffCommands } from './latexdiffCommands';
