--- packages/extension/src/commands/setup/index.ts
@@ -1,4 +1,0 @@
-export {
-  hasAnyUsableSetupCredential,
-  launchSetupAssistant,
-} from './setupAssistantCommand';
