--- packages/extension/src/commands/system/index.ts
@@ -1,11 +1,0 @@
-// Barrel export for system commands
-export {
-  registerMainViewCommands,
-  showImportOptions,
-} from './mainViewCommands';
-export { createSampleProject } from './sampleProjectCommands';
-export { handleTestConnection } from '../tests/connectionTests';
-export {
-  handleTestAgentLoading,
-  handleLoadSpecificAgent,
-} from './yamlCommands';
