--- packages/extension/src/commands/taskFormState/index.ts
@@ -1,2 +1,0 @@
-// Barrel export for main-view task-form state restore commands
-export { registerStateRestoreCommand } from './stateRestoreCommand';
