--- packages/extension/src/frontend/agents/index.ts
@@ -1,6 +1,0 @@
-// Barrel export for frontend agent utilities
-export { agentDirectories } from './AgentDirectoryManager';
-export {
-  getAgentRegistrationSkipReason,
-  promptToAddAgentToConfig,
-} from './register';
