--- packages/extension/src/frontend/files/index.ts
@@ -1,6 +1,0 @@
-// Barrel export for frontend file utilities
-export { FileLister, getFileLister } from './fileLister';
-export {
-  FILE_SELECTION_COMMAND_IDS,
-  MULTIPLE_FILE_COMMANDS,
-} from './fileSelectionRegistry';
