+++ src/tools/agentCliShared.ts
@@ -1,12 +1,12 @@
 // Shared helpers for the agent-CLI tool modules (codex.ts, claudeAgent.ts).
 // Host-agnostic, VS Code-free.
 
-import {
-  registerExecution,
-  releaseOwnedExecutionLeaseAfterFailure,
-} from '@agent/storage';
+import { registerExecution } from '@agent/storage';
 import { type AgentTrace } from '@agent/trace';
-import { captureOwnedExecutionLease } from '@agent/storage/executionLease';
+import {
+  captureOwnedExecutionLease,
+  releaseOwnedExecutionLeaseAfterFailure,
+} from '@agent/storage/executionLease';
 import { AgentCategory } from '@agent/core/definition/AgentDataclass';
 import type { AgentConfig } from '@agent/core/definition/AgentConfig';
 import type { AgentRuntimeHost } from '@agent/runtime/AgentRuntimeHost';
