+++ src/tools/bash.ts
@@ -6,12 +6,12 @@ import {
   finalizeExecution,
   getExecutionStore,
   registerExecution,
-  releaseOwnedExecutionLeaseAfterFailure,
 } from '@agent/storage';
 import {
   captureOwnedExecutionLease,
   markOwnedExecutionLeaseUndurable,
   onOwnedExecutionLeaseLost,
+  releaseOwnedExecutionLeaseAfterFailure,
 } from '@agent/storage/executionLease';
 import {
   TOOL_RESULT_TRUNCATION_HEAD_CHARS,
